Version 1.2.3 imported from tarball.
