+++ lib_src/olcommon/src/debug.F90
@@ -1,0 +1,114 @@
+
+! Copyright 2014 Fabio Cascioli, Jonas Lindert, Philipp Maierhoefer, Stefano Pozzorini
+!
+! This file is part of OpenLoops.
+!
+! OpenLoops is free software: you can redistribute it and/or modify
+! it under the terms of the GNU General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! OpenLoops is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU General Public License for more details.
+!
+! You should have received a copy of the GNU General Public License
+! along with OpenLoops.  If not, see <http://www.gnu.org/licenses/>.
+
+module ol_debug
+  ! precision independent message and error routines
+  use, intrinsic :: iso_c_binding, only:  c_int
+  use, intrinsic :: iso_fortran_env, only : stdout=>output_unit, &
+                                            stderr=>error_unit
+  implicit none
+  private
+  public :: set_verbose, get_verbose, get_error, ol_msg, ol_error, ol_fatal
+  public :: error, verbose, do_not_stop
+
+  integer, save :: verbose = 0
+  integer, save :: error = 0
+  logical, save :: do_not_stop = .false.
+
+  interface ol_msg
+    module procedure ol_print_msg_level, ol_print_msg
+  end interface ol_msg
+
+  interface ol_error
+    module procedure ol_error_msg, ol_error_level
+  end interface ol_error
+
+  contains
+
+  subroutine set_verbose(level)
+    implicit none
+    integer, intent(in) :: level
+    verbose = level
+  end subroutine set_verbose
+
+  subroutine get_verbose(level)
+    implicit none
+    integer, intent(in) :: level
+    verbose = level
+  end subroutine get_verbose
+
+  subroutine ol_print_msg(msg)
+    implicit none
+    character(len=*), intent(in) :: msg
+    call ol_msg(0,msg)
+  end subroutine ol_print_msg
+
+  subroutine ol_print_msg_level(level, msg)
+    implicit none
+    integer, intent(in) :: level
+    character(len=*), intent(in) :: msg
+    if (verbose >= level) write(stdout,*) "[OpenLoops] " // trim(msg)
+  end subroutine ol_print_msg_level
+
+  subroutine ol_error_level(err, msg)
+    implicit none
+    integer, intent(in) :: err
+    character(len=*), intent(in), optional :: msg
+    error = err
+    if (present(msg)) then
+      write(stderr,*) "[OpenLoops] Error: " // trim(msg)
+    end if
+  end subroutine ol_error_level
+
+  subroutine ol_error_msg(msg)
+    implicit none
+    character(len=*), intent(in), optional :: msg
+    call ol_error_level(1,msg)
+  end subroutine ol_error_msg
+
+  function get_error()
+    implicit none
+    integer :: get_error
+    get_error = error
+  end function get_error
+
+  function get_error_c() bind(c,name="ol_get_error")
+    implicit none
+    integer(c_int) :: get_error_c
+    get_error_c = error
+  end function get_error_c
+
+  subroutine ol_fatal(msg, fatal_err)
+    implicit none
+    character(len=*), optional, intent(in) :: msg
+    integer, optional, intent(out) :: fatal_err
+    error = 2
+    if (present(msg)) write(stderr,*) "[OpenLoops] ERROR: " // trim(msg)
+    if (present(fatal_err)) then
+      fatal_err = 1
+    else if (do_not_stop) then
+      if (verbose > 0) write(stderr,*) "[OpenLoops] FATAL ERROR."
+    else
+      if (verbose > 0) then
+        write(stderr,*) "[OpenLoops] STOP."
+      end if
+      stop
+    end if
+  end subroutine ol_fatal
+
+end module ol_debug
