+++ lib_src/openloops/src/contractions.F90
@@ -155,15 +155,15 @@ subroutine cont_EpVVV(B, C, D, Aout)
   complex(REALKIND), intent(out) :: Aout(4)
   complex(REALKIND)              :: C1D4, C1D3, C1D2, C2D3, C2D4, C3D4
   ! all vectors in standard representation
-  ! Aout(1) = B(2) * (C(3)*D(4) - C(4)*D(3)) + B(3) * (C(4)*D(2) - C(2)*D(4)) + B(4) * (C(2)*D(3) - C(3)*D(2))
-  ! Aout(2) = B(3) * (C(4)*D(1) - C(1)*D(4)) + B(4) * (C(1)*D(3) - C(3)*D(1)) + B(1) * (C(3)*D(4) - C(4)*D(3))
-  ! Aout(3) = B(4) * (C(2)*D(1) - C(1)*D(2)) + B(1) * (C(4)*D(2) - C(2)*D(4)) + B(2) * (C(1)*D(4) - C(4)*D(1))
-  ! Aout(4) = B(1) * (C(2)*D(3) - C(3)*D(2)) + B(2) * (C(3)*D(1) - C(1)*D(3)) + B(3) * (C(1)*D(2) - C(2)*D(1))
+  ! aout(1) = b(2) * (c(3)*d(4) - c(4)*d(3)) + b(3) * (c(4)*d(2) - c(2)*d(4)) + b(4) * (c(2)*d(3) - c(3)*d(2))
+  ! aout(2) = b(1) * (c(3)*d(4) - c(4)*d(3)) + b(3) * (c(4)*d(1) - c(1)*d(4)) + b(4) * (c(1)*d(3) - c(3)*d(1))
+  ! aout(3) = b(1) * (c(4)*d(2) - c(2)*d(4)) + b(2) * (c(1)*d(4) - c(4)*d(1)) + b(4) * (c(2)*d(1) - c(1)*d(2))
+  ! aout(4) = b(1) * (c(2)*d(3) - c(3)*d(2)) + b(2) * (c(3)*d(1) - c(1)*d(3)) + b(3) * (c(1)*d(2) - c(2)*d(1))
   ! light-cone representation
-  ! Aout(1) = B(1) * (C(4)*D(3) - C(3)*D(4)) + B(4) * (C(3)*D(1) - C(1)*D(3)) + B(3) * (C(1)*D(4) - C(4)*D(1))
-  ! Aout(2) = B(2) * (C(3)*D(4) - C(4)*D(3)) + B(3) * (C(4)*D(2) - C(2)*D(4)) + B(4) * (C(2)*D(3) - C(3)*D(2))
-  ! Aout(3) = B(3) * (C(1)*D(2) - C(2)*D(1)) + B(1) * (C(2)*D(3) - C(3)*D(2)) + B(2) * (C(3)*D(1) - C(1)*D(3))
-  ! Aout(4) = B(4) * (C(2)*D(1) - C(1)*D(2)) + B(2) * (C(1)*D(4) - C(4)*D(1)) + B(1) * (C(4)*D(2) - C(2)*D(4))
+  ! Aout(1) = i/2 * (B(1) * (C(4)*D(3) - C(3)*D(4)) + B(3) * (C(1)*D(4) - C(4)*D(1)) + B(4) * (C(3)*D(1) - C(1)*D(3)))
+  ! Aout(2) = i/2 * (B(2) * (C(3)*D(4) - C(4)*D(3)) + B(3) * (C(4)*D(2) - C(2)*D(4)) + B(4) * (C(2)*D(3) - C(3)*D(2)))
+  ! Aout(3) = i/2 * (B(1) * (C(2)*D(3) - C(3)*D(2)) + B(2) * (C(3)*D(1) - C(1)*D(3)) + B(3) * (C(1)*D(2) - C(2)*D(1)))
+  ! Aout(4) = i/2 * (B(1) * (C(4)*D(2) - C(2)*D(4)) + B(2) * (C(1)*D(4) - C(4)*D(1)) + B(4) * (C(2)*D(1) - C(1)*D(2)))
   C1D2    = C(1)*D(2) - C(2)*D(1)
   C1D3    = C(1)*D(3) - C(3)*D(1)
   C1D4    = C(1)*D(4) - C(4)*D(1)
