+++ pyol/config/default.cfg
@@ -10,6 +10,10 @@
 
 supported_compilers = gfortran ifort
 fortran_compiler = gfortran
+fortran_tool = auto
+cc = gcc
+cxx = g++
+cpp = cpp
 
 # compiler dependent flags
 gfortran_noautomatic = -fno-automatic
@@ -20,7 +24,7 @@ ifort_noautomatic = -save
 ifort_f77_flags = -132
 ifort_f90_flags =
 ifort_f_flags =
-
+ccflags = -Wall
 
 # generator: 0:off, 1:on, 2:download
 generator = 2
@@ -32,7 +36,7 @@ gjobs = 1
 glog = 0
 
 # release version; max. 8 characters, empty if none
-release =
+release = 1.2.3
 clean =
 
 # precision: [sp (single),] dp (double), [ep (extended, only gfortran),] qp (quad),
@@ -48,7 +52,7 @@ num_jobs = 0
 # compile: 0:off, 1:on (generic libraries only when no process is specified), 2:on
 compile = 2
 # compile extra: compile 'extra' processes (usually extra means real corrections)
-compile_extra = 1
+compile_extra = 0
 
 # compile_libraries/link_libraries: whitespace separated list of rambo, collier, cuttools, samurai
 # if needed, olcommon, oneloop, openloops are added automatically
@@ -67,11 +71,11 @@ debug_flags_1 = -g
 # added to f_flags if debug = 4,5,6,7
 gfortran_debug_flags_4 = -Warray-bounds -fbounds-check
 ifort_debug_flags_4 = -CB
-debug_flags_4 = %%(%(fortran_compiler)s_debug_flags_4)s
+debug_flags_4 = %%(%(fortran_tool)s_debug_flags_4)s
 # Fortran 77 compiler options
-f77_flags = %%(%(fortran_compiler)s_f77_flags)s
+f77_flags = %%(%(fortran_tool)s_f77_flags)s
 # Fortran 90 compiler options
-f90_flags = %%(%(fortran_compiler)s_f90_flags)s
+f90_flags = %%(%(fortran_tool)s_f90_flags)s
 # loop_optimisation is only used for loop_*, virtual_<n>_*, tensorsum_*
 generic_optimisation = -O2
 born_optimisation = -O2
@@ -81,10 +85,10 @@ link_optimisation = -O2
 common_flags =
 link_flags = %(common_flags)s
 # common Fortran compiler options, added to f77_flags and f90_flags
-f_flags = %%(%(fortran_compiler)s_f_flags)s %(common_flags)s
-noautomatic = %%(%(fortran_compiler)s_noautomatic)s
+f_flags = %%(%(fortran_tool)s_f_flags)s %(common_flags)s
+noautomatic = %%(%(fortran_tool)s_noautomatic)s
 
-import_path = 1
+import_env = PATH LD_LIBRARY_PATH CPATH C_INCLUDE_PATH CPLUS_INCLUDE_PATH INTEL_LICENSE_FILE
 math_cmd = math-8.0
 math_flags = -noinit
 lib_src_dir = lib_src
